+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 ### Changes
 - Set the minimum required number of chars to display the emoji suggestions to 2 and made it configurable on Other Settings.
 - Added a setting to toggle whether Enter completes emoji suggestions.
+- Improved Search on the settings so it doesn't need to open the category